update spade code to Nvidia's
